move babel config to a common folder
