+++ package.json
@@ -29,8 +29,5 @@
         "component": "Component"
       }
     }
-  },
-  "devDependencies": {
-    "babel-jest": "^23.6.0"
   }
 }
