+++ packages/x-babel-config/index.js
