+++ packages/x-rollup/package.json
@@ -11,13 +11,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@financial-times/x-babel-config": "file:../x-babel-config",
     "babel-core": "^6.26.3",
     "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
     "chalk": "^2.4.1",
-    "fast-async": "^6.3.7",
     "log-symbols": "^2.2.0",
     "rollup": "^0.63.0",
     "rollup-plugin-babel": "^3.0.7",
